requirements.txt
streamlit
pandas
requests
reportlab
python-dotenv
